polishing tests

+++ tests/tests.py
@@ -5,6 +5,7 @@ import os
 import subprocess
 import tempfile
 
+debug = False
 
 class TestCase(unittest.TestCase):
     def setUp(self):
@@ -24,6 +25,8 @@ class TestCase(unittest.TestCase):
             final_cmd = "%s -- %s" % (self.fcpath, cmd)
         else:
             final_cmd = cmd
+        if debug:
+            print "debug: %r" % (final_cmd,)
         rc = subprocess.call(final_cmd, shell=True)
         self.assertEqual(rc, returncode)
         return True
@@ -37,7 +40,8 @@ def at_most(seconds=None):
             ret = fn(self, *args, **kwargs)
             t1 = time.time()
             td = t1 - t0
-            self.assertLessEqual(td, seconds, "Task took %.1f, not %.1f seconds" % (td, seconds))
+            self.assertLessEqual(td, seconds, "Task took %.1f, not %.1f seconds" \
+                                     % (td, seconds))
             return ret
         return wrapper
     return decorator
@@ -53,15 +57,20 @@ def compile(code=None):
             os.write(fd, code + '\n')
             os.close(fd)
             try:
-                cc_cmd = "%s %s -Os -Wall %s -o %s" % (os.getenv('CC', 'cc'), os.getenv('CFLAGS', ''), source, compiled)
+                cc_cmd = "%s %s -Os -Wall %s -o %s" \
+                    % (os.getenv('CC', 'cc'), os.getenv('CFLAGS', ''),
+                       source, compiled)
                 rc = subprocess.call(cc_cmd, shell=True)
                 self.assertEqual(rc, 0)
 
                 kwargs['compiled'] = compiled
                 ret = fn(self, *args, **kwargs)
             finally:
-                os.unlink(source)
-                os.unlink(compiled)
+                if not debug:
+                    os.unlink(source)
+                    os.unlink(compiled)
+                else:
+                    print "debug: not unlinked %s %s" % (source, compiled)
             return ret
         return wrapper
     return decorator
